requirements.txt
selenium
pytest
pytest-xdist
allure-pytest
requests
python-dotenv
beautifulsoup4
